moved keyboard into its own c source and header file
